controller/requirements.txt
pyyaml
numpy
redis
zerorpc
pygments
Pillow